Using a double to accumulate display time.

+++ src/core/io/display.h
@@ -72,7 +72,8 @@ typedef struct _Display_t {
 
     Program_t programs[Display_Programs_t_CountOf];
     Program_t *active_program;
-    float time;
+
+    double time;
 
     GL_Context_t *context;
     GL_Palette_t palette;
